Add client/chat functionality and node discovery
Replaced the boilerplate code in client.hs with actual client
functionality wherein it discovers the server node given the server
address via a command line argument. It then asks the user to enter a
username and "joins" the chat server. Finally, it spawns a local node
listening for broadcasted messages and loops forever waiting for the
user to send messages to the chat server.

Moved some of the server code to src/Server.hs to clean up the main
section and reduce the imports. Additonally, changed the server main to
register itself such that clients can find it.