Add disconnect handling to server and client
Server now monitors the ports it receives on JoinMessage and correctly
removes said client from the clients map while additionally
broadcasting a message that the client has disconnected.

Client declares a link to the server upon discovering the serverPID but
before joining such that the client will close and necessitate a restart
if the server fails before join or afterwards. I would prefer to do this
with a monitor but I was unable to immediately figure out how.